requirements.txt
# --- API & server ---
fastapi>=0.120.0
uvicorn[standard]==0.35.0
uvloop==0.21.0
httptools==0.6.4
websockets==13.1
httpx==0.28.1

# --- RAG stack ---
chromadb==1.0.20
sentence-transformers==5.1.0
ollama==0.5.3
groq>=0.4.0
redis==5.2.0
rank-bm25==0.2.2  # BM25 keyword search for hybrid retrieval

# --- Core ML deps ---
# Using flexible version for Python 3.12 compatibility
torch>=2.6.0

# --- Security Fixes ---
filelock>=3.20.1  # CVE-2025-68146
starlette>=0.49.1  # CVE-2025-62727
urllib3>=2.6.0  # CVE-2025-50182, CVE-2025-66418
transformers==4.55.2

# --- Data / utils ---
tqdm==4.67.1
pyyaml==6.0.2
pydantic==2.11.7
python-dotenv==1.1.1
python-multipart==0.0.20
nltk==3.9.1  # Natural language toolkit for improved BM25 tokenization

# --- Metrics ---
prometheus-client==0.22.1
prometheus-fastapi-instrumentator==7.1.0

# --- Analysis & plotting ---
pandas==2.3.2
matplotlib==3.10.6

# --- Testing (needed for async tests) ---
pytest-asyncio>=0.23.0
